Share validation tests
