+++ project/plugins.sbt
@@ -5,7 +5,7 @@ resolvers ++= Resolver.sonatypeOssRepos("snapshots")
 val crossVer = "1.3.2"
 val scalaJSVersion = "1.17.0"
 val scalaNativeVersion = "0.5.5"
-val scalafix = "0.12.1"
+val scalafix = "0.13.0"
 
 // includes sbt-dynver sbt-pgp sbt-sonatype sbt-git
 addSbtPlugin("com.github.sbt" % "sbt-ci-release" % "1.6.1")
