+++ sconfig/js/src/main/scala/org/ekrich/config/parser/PlatformConfigDocumentFactory.scala
@@ -1,7 +1,7 @@
 package org.ekrich.config.parser
 
 /**
- * [[ConfigFactoryDocument]] methods for Scala.js platform
+ * [[ConfigDocumentFactory]] methods for Scala.js platform
  */
 abstract class PlatformConfigDocumentFactory
     extends ConfigDocumentFactoryShared {}
