+++ sconfig/jvm-native/src/test/scala/org/ekrich/config/impl/ConfigFactoryDocumentTest.scala
@@ -3,6 +3,6 @@ package org.ekrich.config.impl
 // import org.junit.Assert._
 // import org.junit.Test
 
-class ConfigFactoryDocumentTest {
+class ConfigDocumentFactoryTest {
   // Empty for now - Scala.js has parody with Scala Native for now
 }
