+++ sconfig/jvm/src/main/scala/org/ekrich/config/parser/PlatformConfigDocumentFactory.scala
@@ -1,7 +1,7 @@
 package org.ekrich.config.parser
 
 /**
- * [[ConfigFactoryDocument]] methods for Scala JVM platform
+ * [[ConfigDocumentFactory]] methods for Scala JVM platform
  */
 abstract class PlatformConfigDocumentFactory
     extends ConfigDocumentFactoryJvmNative {}
