+++ sconfig/jvm/src/test/scala/org/ekrich/config/impl/TestPath.scala
@@ -1,0 +1,8 @@
+package org.ekrich.config.impl
+
+import java.io.File
+
+// See: https://github.com/scala-native/scala-native/issues/4077
+object TestPath {
+  def file(): File = new File("src/test/resources")
+}
