+++ sconfig/native/src/main/scala/org/ekrich/config/parser/PlatformConfigDocumentFactory.scala
@@ -1,7 +1,7 @@
 package org.ekrich.config.parser
 
 /**
- * [[ConfigFactoryDocument]] methods for Scala Native platform
+ * [[ConfigDocumentFactory]] methods for Scala Native platform
  */
 abstract class PlatformConfigDocumentFactory
     extends ConfigDocumentFactoryJvmNative {}
