+++ sconfig/native/src/test/resources/validate-invalid.conf
@@ -1,0 +1,30 @@
+string1="a string"
+string2=107
+string3={ a : b }
+string4=[]
+int1=203
+int2="foo"
+int3={ q : s }
+float1="the string"
+float2=false
+float3=[ 4, 5, 6 ]
+bool1=709
+bool2="string!"
+bool3={}
+null1=10000
+null2="hello world"
+null3=true
+object1={ z : s }
+object2=[]
+object3=12345
+array1=[1,2,"foo"]
+array2=[7,8,9]
+array3=[{ n : m }, 10]
+array4=[42, 43]
+array5=64
+emptyArray1=[1,2,3]
+emptyArray2=["a","b","c"]
+
+a.b.c.d.e.f.g = 100
+a.b.c.d.e.f.h = "foo"
+a.b.c.d.e.f.i = []
