+++ sconfig/native/src/test/resources/validate-reference.conf
@@ -1,0 +1,32 @@
+string1="foo"
+string2="bar"
+string3="baz"
+string4="hello"
+int1=10
+int2=11
+int3=12
+float1=3.14
+float2=3.2
+float3=3.3
+bool1=true
+bool2=false
+bool3=true
+null1=null
+null2=null
+null3=null
+object1={ a : b }
+object2={ c : d }
+object3={ e : f }
+array1=[1,2,3]
+array2=[a,b,c]
+array3=[true, true, false]
+array4=[{}, {}]
+array5=[]
+emptyArray1=[]
+emptyArray2=[]
+willBeMissing=90009
+
+a.b.c.d.e.f.g = true
+a.b.c.d.e.f.h = true
+a.b.c.d.e.f.i = true
+a.b.c.d.e.f.j = true
